+++ .flake8
@@ -8,3 +8,6 @@ ignore =
 
 exclude =
     .venv/**
+
+per-file-ignores =
+    **/**/__init__.py:F401
